fix: clippy cleanup for 3.7.0 release
- Remove unused HashMap import from btree.rs (converted to LruCache)
- Fix unnecessary u64 cast in async_io/coordinator.rs
- Collapse nested if in backend.rs ef_search validation
- Add module-level clippy allows with documented reasons on V3 files
  (type_complexity, collapsible_if, items_after_test_module,
  manual_async_fn) for patterns inherent to the V3 design
- Fix &PathBuf to &Path in edge_compat.rs
- Remove unused SqliteGraphError import from v3_query_truth_tests.rs
- Remove duplicate Path import in edge_compat test module

+++ sqlitegraph-core/src/backend/native/v3/async_io/coordinator.rs
@@ -51,7 +51,7 @@ impl AsyncFileCoordinator {
         if page_id == 0 {
             0
         } else {
-            V3_HEADER_SIZE + (page_id - 1) * DEFAULT_PAGE_SIZE as u64
+            V3_HEADER_SIZE + (page_id - 1) * DEFAULT_PAGE_SIZE
         }
     }
